+++ h3/sql/updates/h3--3.7.1--3.7.2.sql
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Bytes & Brains
+ * Copyright 2022 Bytes & Brains
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,4 +15,6 @@
  */
 
 -- complain if script is sourced in psql, rather than via CREATE EXTENSION
-\echo Use "ALTER EXTENSION h3 UPDATE TO 'unreleased'" to load this file. \quit
+\echo Use "ALTER EXTENSION h3 UPDATE TO '3.7.2'" to load this file. \quit
+
+-- no changes